where command - prints student's current service

+++ src/System/Location.java
@@ -3,6 +3,6 @@ package System;
 import java.io.Serializable;
 
 public interface Location extends Serializable {
-    public double getLatitude();
-    public double getLongitude();
+    public long getLatitude();
+    public long getLongitude();
 }
